refactor: move GET parameter handling mostly into its own file js/browser/params.ts and gate WebGL renderer behind 'gpu' flag
